fix: resolve all compiler warnings across debug/release/native presets
- benchmarks/bench_kernels.cpp: add (void) casts for [[nodiscard]] ops
  discarded inside timing lambdas
- chapters/ch14_inference/main.cpp: fix implicit int→float conversions in
  percentage calculations (4 sites)
- chapters/ch21_math_neurons/main.cpp: fix sign-conversion warnings —
  change loop variables and std::min calls from int to std::size_t (8 sites);
  change op_idx from int to std::size_t for std::array indexing
- tests/test_ch25_long_context.cpp: add (void) casts for [[nodiscard]]
  forward_one() results discarded in test setup

All 474 tests pass; zero warnings on debug, release, and native presets.

+++ chapters/ch14_inference/main.cpp
@@ -46,9 +46,9 @@ static void section_temperature() {
 
         std::cout << std::format("  {:>8.1f}  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%\n",
                                   temp,
-                                  100.0f * counts[0] / N, 100.0f * counts[1] / N,
-                                  100.0f * counts[2] / N, 100.0f * counts[3] / N,
-                                  100.0f * counts[4] / N);
+                                  100.0f * static_cast<float>(counts[0]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[1]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[2]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[3]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[4]) / static_cast<float>(N));
     }
 
     std::cout << "  Low temperature concentrates mass on the peak; high flattens it.\n";
@@ -77,9 +77,9 @@ static void section_top_k() {
         std::cout << std::format("  {:>4}  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%"
                                   "  [top-{}]\n",
                                   k,
-                                  100.0f * counts[0] / N, 100.0f * counts[1] / N,
-                                  100.0f * counts[2] / N, 100.0f * counts[3] / N,
-                                  100.0f * counts[4] / N, k);
+                                  100.0f * static_cast<float>(counts[0]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[1]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[2]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[3]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[4]) / static_cast<float>(N), k);
     }
 
     std::cout << "  Top-k prevents probability mass leaking to very low-probability tokens.\n";
@@ -106,9 +106,9 @@ static void section_top_p() {
 
         std::cout << std::format("  {:>6.2f}  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%  {:>7.1f}%\n",
                                   p,
-                                  100.0f * counts[0] / N, 100.0f * counts[1] / N,
-                                  100.0f * counts[2] / N, 100.0f * counts[3] / N,
-                                  100.0f * counts[4] / N);
+                                  100.0f * static_cast<float>(counts[0]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[1]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[2]) / static_cast<float>(N), 100.0f * static_cast<float>(counts[3]) / static_cast<float>(N),
+                                  100.0f * static_cast<float>(counts[4]) / static_cast<float>(N));
     }
 
     std::cout << "  Nucleus sampling adapts the candidate set to the distribution shape.\n";
@@ -197,7 +197,7 @@ static void section_comparison() {
         std::cout << std::format("  {:>20}: ", label);
         for (int tok = 0; tok < 5; ++tok)
             std::cout << std::format("  tok{}={:.1f}%", tok,
-                                      100.0f * counts[tok] / N);
+                                      100.0f * static_cast<float>(counts[tok]) / static_cast<float>(N));
         std::cout << '\n';
     };
